ClassIncomplete for binding incomplete classes
Incomplete classes would error with bbl::Class because Class would try
and statically grab layout and type_traits info about the type, which
would fail in the incomplete case.

Rather than complicate bbl::Class we add a different type to bind them.

Note that it's not possible to handle this transparenty, because trying
to guess whether a type is incomplete or not in TMP is undefined
behaviour.