Add support for inheriting in C
Generates a trampoline class that takes a virtual function table
in the form of a list of function pointers.

Classes defined in bblext namespace can be bound and are pasted
into the binding library. This allows exposing protected methods
to the C API.

first pass inheritance working

add bblext class pasting